# src/swagger_output.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "v1.0.0",
    "title": "Lend A Friend",
    "description": "Lend A Friend Mobile app REST API server"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": ""
    }
  ],
  "paths": {
...
    "/api/users/change-password": {
      "post": {
        "operationId": "UsersController.changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "description": "",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Change password",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/get-user": {
      "post": {
        "operationId": "UsersController.getUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "description": "",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Get user",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/login": {
      "post": {
        "operationId": "UsersController.login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "description": "",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Login",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/register": {
      "post": {
        "operationId": "UsersController.register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "description": "",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Register",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/send-verification-code": {
      "post": {
        "operationId": "UsersController.sendVerificationCode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "description": "",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Send verification code",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/update-password": {
      "post": {
        "operationId": "UsersController.updatePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "description": "",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Update password",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/validate-email": {
      "post": {
        "operationId": "UsersController.validateEmail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "description": "",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Validate email",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/verify-email": {
      "post": {
        "operationId": "UsersController.verifyEmail",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "description": "",
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            },
            "description": "Successful response"
          }
        },
        "summary": "Verify email",
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  }
}